+++ docs/changelog.md
@@ -13,3 +13,4 @@ Version 1.1.0 ()
 * Added Template mapping closure in `app/_configs/default.php` (`router.template`).
 * Added docs for Configuration and Models.
 * Removed `composer.lock`.
+* Added `ini_set('display_errors', 'on');`